+++ Assets/Tests/InputSystem/APIVerificationTests.cs
@@ -794,6 +794,10 @@ class APIVerificationTests
     [Property("Exclusions", @"1.0.0
         public static event System.Action<UnityEngine.InputSystem.LowLevel.InputEventPtr, UnityEngine.InputSystem.InputDevice> onEvent;
     ")]
+    // Mouse and Touchscreen implement internal IEventMerger interface
+    [Property("Exclusions", @"1.0.0
+        public class Touchscreen : UnityEngine.InputSystem.Pointer, UnityEngine.InputSystem.LowLevel.IInputStateCallbackReceiver
+    ")]
     [ScopedExclusionProperty("1.0.0", "UnityEngine.InputSystem.Editor", "public sealed class InputControlPathEditor : System.IDisposable", "public void OnGUI(UnityEngine.Rect rect);")]
     public void API_MinorVersionsHaveNoBreakingChanges()
     {