+++ Packages/com.unity.inputsystem/InputSystem/Devices/InputDevice.cs
@@ -620,14 +620,15 @@ namespace UnityEngine.InputSystem
             HasStateCallbacks = 1 << 1,
             HasControlsWithDefaultState = 1 << 2,
             HasDontResetControls = 1 << 3,
+            HasEventMerger = 1 << 4,
 
-            Remote = 1 << 4, // It's a local mirror of a device from a remote player connection.
-            Native = 1 << 5, // It's a device created from data surfaced by NativeInputRuntime.
+            Remote = 1 << 5, // It's a local mirror of a device from a remote player connection.
+            Native = 1 << 6, // It's a device created from data surfaced by NativeInputRuntime.
 
-            DisabledInFrontend = 1 << 6, // Explicitly disabled on the managed side.
-            DisabledInRuntime = 1 << 7, // Disabled in the native runtime.
-            DisabledWhileInBackground = 1 << 8, // Disabled while the player is running in the background.
-            DisabledStateHasBeenQueriedFromRuntime = 1 << 9, // Whether we have fetched the current enable/disable state from the runtime.
+            DisabledInFrontend = 1 << 7, // Explicitly disabled on the managed side.
+            DisabledInRuntime = 1 << 8, // Disabled in the native runtime.
+            DisabledWhileInBackground = 1 << 9, // Disabled while the player is running in the background.
+            DisabledStateHasBeenQueriedFromRuntime = 1 << 10, // Whether we have fetched the current enable/disable state from the runtime.
 
             CanRunInBackground = 1 << 11,
             CanRunInBackgroundHasBeenQueried = 1 << 12,
@@ -756,7 +757,7 @@ namespace UnityEngine.InputSystem
 
         internal bool hasDontResetControls
         {
-            get => (m_DeviceFlags & DeviceFlags.HasDontResetControls) != 0;
+            get => (m_DeviceFlags & DeviceFlags.HasDontResetControls) == DeviceFlags.HasDontResetControls;
             set
             {
                 if (value)
@@ -768,13 +769,25 @@ namespace UnityEngine.InputSystem
 
         internal bool hasStateCallbacks
         {
-            get => (m_DeviceFlags & DeviceFlags.HasStateCallbacks) != 0;
+            get => (m_DeviceFlags & DeviceFlags.HasStateCallbacks) == DeviceFlags.HasStateCallbacks;
             set
             {
                 if (value)
                     m_DeviceFlags |= DeviceFlags.HasStateCallbacks;
                 else
                     m_DeviceFlags &= ~DeviceFlags.HasStateCallbacks;
+            }
+        }
+
+        internal bool hasEventMerger
+        {
+            get => (m_DeviceFlags & DeviceFlags.HasEventMerger) == DeviceFlags.HasEventMerger;
+            set
+            {
+                if (value)
+                    m_DeviceFlags |= DeviceFlags.HasEventMerger;
+                else
+                    m_DeviceFlags &= ~DeviceFlags.HasEventMerger;
             }
         }