+++ Packages/com.unity.inputsystem/InputSystem/Devices/Precompiled/FastMouse.partial.cs
@@ -49,7 +49,7 @@ namespace UnityEngine.InputSystem
             OnStateEvent(eventPtr);
         }
 
-        public unsafe bool MergeForward(InputEventPtr currentEventPtr, InputEventPtr nextEventPtr)
+        internal static unsafe bool MergeForward(InputEventPtr currentEventPtr, InputEventPtr nextEventPtr)
         {
             if (currentEventPtr.type != StateEvent.Type || nextEventPtr.type != StateEvent.Type)
                 return false;
@@ -70,6 +70,11 @@ namespace UnityEngine.InputSystem
             nextState->delta += currentState->delta;
             nextState->scroll += currentState->scroll;
             return true;
+        }
+
+        bool IEventMerger.MergeForward(InputEventPtr currentEventPtr, InputEventPtr nextEventPtr)
+        {
+            return MergeForward(currentEventPtr, nextEventPtr);
         }
     }
 }