+++ Packages/com.unity.inputsystem/InputSystem/InputManager.cs
@@ -1143,6 +1143,10 @@ namespace UnityEngine.InputSystem
                 m_HaveDevicesWithStateCallbackReceivers = true;
             }
 
+            // If the device has event merger, make a note of it.
+            if (device is IEventMerger)
+                device.hasEventMerger = true;
+
             // If the device wants before-render updates, enable them if they
             // aren't already.
             if (device.updateBeforeRender)
@@ -2518,7 +2522,7 @@ namespace UnityEngine.InputSystem
                 for (var i = 0; i < m_DevicesCount; ++i)
                 {
                     var device = m_Devices[i];
-                    if ((device.m_DeviceFlags & InputDevice.DeviceFlags.HasStateCallbacks) == 0)
+                    if (!device.hasStateCallbacks)
                         continue;
 
                     // NOTE: We do *not* perform a buffer flip here as we do not want to change what is the
@@ -2951,6 +2955,8 @@ namespace UnityEngine.InputSystem
             m_InputEventStream = new InputEventStream(ref eventBuffer, m_Settings.maxQueuedEventsPerUpdate);
             var totalEventBytesProcessed = 0U;
 
+            InputEvent* skipEventMergingFor = null;
+
             // Handle events.
             while (m_InputEventStream.remainingEventCount > 0)
             {
@@ -3087,7 +3093,7 @@ namespace UnityEngine.InputSystem
                 }
 
                 // Check if the device wants to merge successive events.
-                if (!settings.disableRedundantEventsMerging && device is IEventMerger merger)
+                if (!skipEvent && !settings.disableRedundantEventsMerging && device.hasEventMerger && currentEventReadPtr != skipEventMergingFor)
                 {
                     // NOTE: This relies on events in the buffer being consecutive for the same device. This is not
                     //       necessarily the case for events coming in from the background event queue where parallel
@@ -3095,11 +3101,50 @@ namespace UnityEngine.InputSystem
                     //       new buffering scheme for input events working in the native runtime.
 
                     var nextEvent = m_InputEventStream.Peek();
-                    if (nextEvent != null && merger.MergeForward(currentEventReadPtr, nextEvent))
+                    if (nextEvent != null && currentEventReadPtr->deviceId == nextEvent->deviceId)
                     {
-                        // Event was merged into next event, skipping.
-                        m_InputEventStream.Advance(leaveEventInBuffer: false);
-                        continue;
+                        if (((IEventMerger)device).MergeForward(currentEventReadPtr, nextEvent))
+                        {
+                            // Event was merged into next event, skipping.
+                            m_InputEventStream.Advance(leaveEventInBuffer: false);
+                            continue;
+                        }
+
+                        // If we can't merge current event with next one for any reason, we assume the next event
+                        // carries crucial entropy (button changed state, phase changed, counter changed, etc).
+                        // Hence semantic meaning for current event is "can't merge current with next because next is different".
+                        // But semantic meaning for next event is "next event carries important information and should be preserved",
+                        // from that point of few next event should not be merged with current nor with _next after next_ event.
+                        //
+                        // For example, given such stream of events:
+                        // Mouse       Mouse       Mouse       Mouse       Mouse       Mouse       Mouse
+                        // Event no1   Event no2   Event no3   Event no4   Event no5   Event no6   Event no7
+                        // Time 1      Time 2      Time 3      Time 4      Time 5      Time 6      Time 7
+                        // Pos(10,20)  Pos(12,21)  Pos(13,23)  Pos(14,24)  Pos(16,25)  Pos(17,27)  Pos(18,28)
+                        // Delta(1,1)  Delta(2,1)  Delta(1,2)  Delta(1,1)  Delta(2,1)  Delta(1,2)  Delta(1,1)
+                        // BtnLeft(0)  BtnLeft(0)  BtnLeft(0)  BtnLeft(1)  BtnLeft(1)  BtnLeft(1)  BtnLeft(1)
+                        //
+                        // if we then merge without skipping next event here:
+                        //                         Mouse                                           Mouse
+                        //                         Event no3                                       Event no7
+                        //                         Time 3                                          Time 7
+                        //                         Pos(13,23)                                      Pos(18,28)
+                        //                         Delta(4,4)                                      Delta(5,5)
+                        //                         BtnLeft(0)                                      BtnLeft(1)
+                        //
+                        // As you can see, the event no4 containing mouse button press was lost,
+                        // and with it we lose the important information of timestamp of mouse button press.
+                        //
+                        // With skipping merging next event we will get:
+                        //                         Mouse       Mouse                               Mouse
+                        //                         Time 3      Time 4                              Time 7
+                        //                         Event no3   Event no4                           Event no7
+                        //                         Pos(13,23)  Pos(14,24)                          Pos(18,28)
+                        //                         Delta(3,3)  Delta(1,1)                          Delta(4,4)
+                        //                         BtnLeft(0)  BtnLeft(1)                          BtnLeft(1)
+                        //
+                        // And no4 is preserved, with the exact timestamp of button press.
+                        skipEventMergingFor = nextEvent;
                     }
                 }
 
@@ -3148,8 +3193,7 @@ namespace UnityEngine.InputSystem
                         //       a global ordering of events as there may be multiple substreams (e.g. each individual touch)
                         //       that are generated in the backend and would require considerable work to ensure monotonically
                         //       increasing timestamps across all such streams.
-                        var deviceIsStateCallbackReceiver = (device.m_DeviceFlags & InputDevice.DeviceFlags.HasStateCallbacks) ==
-                            InputDevice.DeviceFlags.HasStateCallbacks;
+                        var deviceIsStateCallbackReceiver = device.hasStateCallbacks;
                         if (currentEventTimeInternal < device.m_LastUpdateTimeInternal &&
                             !(deviceIsStateCallbackReceiver && device.stateBlock.format != eventPtr.stateFormat))
                         {